humidity is now on the bed page
