+++ lib/visual_garden_web/live/product_live/index.ex
@@ -1,4 +1,6 @@
 defmodule VisualGardenWeb.ProductLive.Index do
+  alias VisualGarden.Repo
+  alias VisualGardenWeb.HomeBadge
   alias VisualGarden.MyDateTime
   alias VisualGarden.Planner
   alias VisualGarden.Authorization.UnauthorizedError
@@ -164,6 +166,48 @@ defmodule VisualGardenWeb.ProductLive.Index do
     {:ok, _} = Gardens.delete_product(product)
 
     {:noreply, assign_products(socket)}
+  end
+
+  @impl true
+  def handle_event("water", %{"bed_id" => bid}, socket) do
+    Repo.transaction(fn ->
+      bed = Gardens.get_product!(bid)
+      garden = Gardens.get_garden!(bed.garden_id)
+      Authorization.authorize_garden_modify(garden.id, socket.assigns.current_user)
+
+      {:ok, _} =
+        Gardens.create_event_log("water", %{
+          "event_time" => MyDateTime.utc_now(),
+          "product_id" => bid
+        })
+    end)
+
+    {:noreply,
+     HomeBadge.badge_socket(socket)
+     |> assign_products()
+     |> assign_actions()
+     |> push_patch(to: ~p"/gardens/#{socket.assigns.garden.id}/beds")}
+  end
+
+  @impl true
+  def handle_event("humidity", %{"bed_id" => bid}, socket) do
+    Repo.transaction(fn ->
+      bed = Gardens.get_product!(bid)
+      garden = Gardens.get_garden!(bed.garden_id)
+      Authorization.authorize_garden_modify(garden.id, socket.assigns.current_user)
+
+      {:ok, _} =
+        Gardens.create_event_log("humidity", %{
+          "event_time" => MyDateTime.utc_now(),
+          "product_id" => bid
+        })
+    end)
+
+    {:noreply,
+     HomeBadge.badge_socket(socket)
+     |> assign_products()
+     |> assign_actions()
+     |> push_patch(to: ~p"/gardens/#{socket.assigns.garden.id}/beds")}
   end
 
   def friendly_type(name) do
